+++ liana/tests/test_causalnet.py
@@ -25,8 +25,7 @@ def test_caulsalnet():
                                      input_scores, 
                                      output_scores, 
                                      node_weights=node_weights,
-                                     verbose=False,
-                                     show_solver_output=False
+                                     verbose=False
                                      )
 
     assert problem.weights == [1.0, 0.01, 1.0]
@@ -43,8 +42,7 @@ def test_causalnet_noweights():
                                      input_scores,
                                      output_scores,
                                      node_weights={"N1": 1, "N2": 0},
-                                     verbose=False,
-                                     show_solver_output=False
+                                     verbose=False
                                      )
     assert df_res['source_pred_val'].values.sum() == 9
     assert df_res['target_pred_val'].values.sum() == 10